docs: refactoring batch operations async first
